Rename to CountAllProjects and ListAllProjects refactor
