+++ src/Api/Dfe.Complete.Api/Controllers/ProjectsController.cs
@@ -6,7 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using Swashbuckle.AspNetCore.Annotations;
 using Dfe.Complete.Application.Projects.Commands.CreateProject;
 using Dfe.Complete.Application.Projects.Model;
-using Dfe.Complete.Application.Projects.Queries.CountProjects;
+using Dfe.Complete.Application.Projects.Queries.CountAllProjects;
 using Dfe.Complete.Domain.Entities;
 using Dfe.Complete.Application.Projects.Queries.GetProject;
 using Dfe.Complete.Application.Projects.Queries.ListAllProjects;
