+++ src/Core/Dfe.Complete.Application/Projects/Model/ListAllProjectsResultModel.cs
@@ -1,5 +1,4 @@
-﻿using Dfe.Complete.Domain.Entities;
-using Dfe.Complete.Domain.Enums;
+﻿using Dfe.Complete.Domain.Enums;
 using Dfe.Complete.Domain.ValueObjects;
 
 namespace Dfe.Complete.Application.Projects.Model;
